requirements.txt
gradio>=5.15
Pillow
